requirements.txt
playwright>=1.40.0
pynput>=1.7.6
pyautogui>=0.9.54
apscheduler>=3.10.4
gspread>=6.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
requests>=2.31.0
pystray>=0.19.5
Pillow>=10.2.0
pyinstaller>=6.3.0
win10toast>=0.9
